refactor(pipeline-cli): dedup ci-required + gh-phoenix to one core each (#2442)
Both `ci-required` and `gh-phoenix` carried a git-tracked standalone package
AND an un-delegated, drift-prone copy inside `pipeline-cli/src/tools/` with no
shared core. The `gh-phoenix` copies had already diverged (pipeline-cli's lint
carried the #1766 frontmatter check, the standalone did not) — the exact silent
drift on §CP gate logic this closes.

Consolidate each tool to a single source of truth the standalone package owns,
imported by the pipeline-cli tool:

- ci-required: pipeline-cli's `ci-required` tool now imports the verdict core
  from `@kampus/ci-required`; the copied core + its unit test are removed. The
  standalone package (and its CI bin `node packages/ci-required/src/bin.ts`) is
  retained unchanged.
- gh-phoenix: promote the advanced lint core (gh-call grep #743 + strict-YAML
  frontmatter #1766) into the standalone `@kampus/gh-phoenix`, which the live
  `gh` REST shim on `.claude/settings.json` PATH already consumes; pipeline-cli's
  `gh-phoenix lint-skills` now imports that core. The copied lint/resolve/router
  cores + their unit tests are removed. The standalone `bin.ts lint-skills` is
  updated to honor the frontmatter check so both IO shells behave identically.

`@kampus/ci-required` + `@kampus/gh-phoenix` added to pipeline-cli deps as
`workspace:*` (the internal-dep convention); `yaml` (catalog:) added to
gh-phoenix for the frontmatter parse. No behavior change on any consumed
surface: the `ci-required` CI bin and the `skill-gh-lint` gate run identically.
