+++ packages/gh-phoenix/package.json
@@ -18,7 +18,8 @@
 	},
 	"dependencies": {
 		"@effect/platform-node": "catalog:",
-		"effect": "catalog:"
+		"effect": "catalog:",
+		"yaml": "catalog:"
 	},
 	"devDependencies": {
 		"@effect/tsgo": "catalog:",
