+++ packages/gh-phoenix/src/bin.ts
@@ -3,9 +3,10 @@
  * `gh-phoenix` CLI — the operable surface for issue #743. Two roles, dispatched
  * on the first argument:
  *
- *   - `gh-phoenix lint-skills <file>...` — run the skill grep-lint (the pure
- *     `lintCorpus` core) over the handed files, emit the scanned scope, and FAIL
- *     CLOSED (exit 3) on zero scope per ADR 0092; exit 2 on any finding, 0 clean.
+ *   - `gh-phoenix lint-skills <file>...` — run the skill lint (the pure `lintCorpus`
+ *     core: GraphQL-path `gh` calls + invalid YAML frontmatter, #743/#1766) over the
+ *     handed files, emit each check's scanned scope, and FAIL CLOSED (exit 3) on zero
+ *     scope in either check per ADR 0092; exit 2 on any finding, 0 clean.
  *
  *   - `gh-phoenix <any other gh args>` — the `gh` SHIM. When this binary shadows
  *     `gh` on the subagent PATH (as `gh`, symlinked/wrapped to here), it routes
@@ -86,7 +87,9 @@ const readFileOrSkip = (file: string): string | null => {
 
 const fileArg = Argument.string("file").pipe(
 	Argument.atLeast(1),
-	Argument.withDescription("one or more skill-corpus file paths to lint for GraphQL-path gh calls"),
+	Argument.withDescription(
+		"one or more skill-corpus file paths to lint for GraphQL-path gh calls and invalid YAML frontmatter",
+	),
 );
 
 const lintSkills = Command.make(
@@ -101,14 +104,19 @@ const lintSkills = Command.make(
 
 		const result = lintCorpus(scanInput);
 
-		// ADR 0092: emit what was scanned (scope is observable), THEN judge.
+		// ADR 0092: emit what each check scanned (scope is observable), THEN judge.
 		yield* Console.log(
-			`gh-phoenix lint-skills: scanned ${result.scanned.length} file(s)` +
+			`gh-phoenix lint-skills: gh-call scan scanned ${result.scanned.length} file(s)` +
 				(result.scanned.length > 0 ? `:` : ` (zero scope)`),
 		);
 		for (const f of result.scanned) yield* Console.log(`  scanned: ${f}`);
+		yield* Console.log(
+			`gh-phoenix lint-skills: frontmatter check scanned ${result.frontmatterScanned.length} file(s)` +
+				(result.frontmatterScanned.length > 0 ? `:` : ` (zero scope)`),
+		);
+		for (const f of result.frontmatterScanned) yield* Console.log(`  frontmatter-scanned: ${f}`);
 
-		// ADR 0092: zero scope is a FAIL, never a silent PASS.
+		// ADR 0092: zero scope in EITHER check is a FAIL, never a silent PASS.
 		if (isZeroScope(result)) {
 			yield* Console.error(
 				"gh-phoenix lint-skills: FAIL — scanned zero files (zero-scope fail-closed, ADR 0092).",
@@ -116,20 +124,33 @@ const lintSkills = Command.make(
 			return yield* Effect.fail(new ZeroScope());
 		}
 
-		if (result.findings.length === 0) {
+		const total = result.findings.length + result.frontmatterFindings.length;
+		if (total === 0) {
 			yield* Console.log(
-				"gh-phoenix lint-skills: clean — no GraphQL-path gh calls in the scanned skill corpus.",
+				"gh-phoenix lint-skills: clean — no GraphQL-path gh calls and all frontmatter parses as strict YAML.",
 			);
 			return;
 		}
 
-		yield* Console.error(
-			`gh-phoenix lint-skills: FAIL — ${result.findings.length} GraphQL-path gh call(s) in the skill corpus (REST-only on this org, #743):`,
-		);
-		for (const f of result.findings) {
-			yield* Console.error(`  ${f.file}:${f.line}: ${f.matched} — ${f.reason}`);
+		if (result.findings.length > 0) {
+			yield* Console.error(
+				`gh-phoenix lint-skills: FAIL — ${result.findings.length} GraphQL-path gh call(s) in the skill corpus (REST-only on this org, #743):`,
+			);
+			for (const f of result.findings) {
+				yield* Console.error(`  ${f.file}:${f.line}: ${f.matched} — ${f.reason}`);
+			}
 		}
-		return yield* Effect.fail(new FindingsFound({count: result.findings.length}));
+
+		if (result.frontmatterFindings.length > 0) {
+			yield* Console.error(
+				`gh-phoenix lint-skills: FAIL — ${result.frontmatterFindings.length} file(s) with invalid YAML frontmatter (a mid-sentence colon-space in an unquoted scalar reparses as a mapping; quote the value or use a block scalar — #1766):`,
+			);
+			for (const f of result.frontmatterFindings) {
+				yield* Console.error(`  ${f.file}: ${f.reason}`);
+			}
+		}
+
+		return yield* Effect.fail(new FindingsFound({count: total}));
 	}),
 ).pipe(
 	Command.withDescription(
