+++ packages/pipeline-cli/package.json
@@ -36,6 +36,8 @@
 	},
 	"dependencies": {
 		"@effect/platform-node": "catalog:",
+		"@kampus/ci-required": "workspace:*",
+		"@kampus/gh-phoenix": "workspace:*",
 		"effect": "catalog:",
 		"fast-xml-parser": "catalog:",
 		"yaml": "catalog:"
